[refactor] dto location
