+++ back/src/main/java/toy/bookchat/bookchat/domain/book/Book.java
@@ -1,16 +1,15 @@
-package toy.bookchat.bookchat.domain.book.model;
+package toy.bookchat.bookchat.domain.book;
 
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import javax.persistence.*;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import toy.bookchat.bookchat.domain.bookshelf.BookShelf;
 
 @Entity
 @Getter
@@ -28,4 +27,6 @@ public class Book {
     private List<String> authors = new ArrayList<>();
     private String publisher;
     private String bookCoverImageUrl;
+    @OneToMany
+    private List<BookShelf> bookShelves = new ArrayList<>();
 }
