+++ back/src/main/java/toy/bookchat/bookchat/domain/book/repository/BookRepository.java
@@ -2,7 +2,7 @@ package toy.bookchat.bookchat.domain.book.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import toy.bookchat.bookchat.domain.book.model.Book;
+import toy.bookchat.bookchat.domain.book.Book;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
