+++ back/src/main/java/toy/bookchat/bookchat/domain/user/api/UserController.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import toy.bookchat.bookchat.domain.user.dto.UserProfileResponse;
+import toy.bookchat.bookchat.domain.user.api.dto.UserProfileResponse;
 import toy.bookchat.bookchat.security.user.UserPrincipal;
 
 @RestController
